separate some methods and modules
