+++ ios_tests/lib/ios/specs/ios/command/pasteboard.rb
@@ -1,6 +1,8 @@
 # rake ios[common/pasteboard]
 describe 'ios/command/pasteboard' do
   t 'pasteboard' do
+    # set blank before testing because pasteboard is remaining during launching simulators
+    set_pasteboard content: ''
     get_pasteboard.must_be_empty
 
     set_pasteboard content: 'sample content'
